using size_t more toward being in compliance with INT01-C, however it doesn't always seem appropriate to use it, and I kind of prefer using uint32_t or uint64_t, in some instances uint is required and therefore not replaced with size_t, need to think more about how strictly to comply with INT01-C

+++ source/datacontainer.c
@@ -19,7 +19,7 @@ static int destroyDataContainer(dataContainer **thisPointer);
  * the data pointer is of type unsigned char*, pointing to up to 2^64 bytes 
  *
  */
-dataContainer* newDataContainer(uint64_t bytesize)
+dataContainer* newDataContainer(size_t bytesize)
 {
   dataContainer *this;
